Add Keyword domain object. Move status logic to service

+++ grails-app/domain/seminario/domain/Status.groovy
@@ -2,10 +2,10 @@ package seminario.domain
 
 class Status {
 
-    Integer accuracy
+    Float accuracy
     String prediction
 
-    Status(Integer accuracy, String prediction) {
+    Status(Float accuracy, String prediction) {
         this.accuracy = accuracy
         this.prediction = prediction
     }
